refactor(gateway): extract SyncCoordinator from GatewayRepository (#106 phase 1)
First half of the #106 split. Pulls the script-registration + BALANCED
filter cluster out of GatewayRepository into its own @Singleton
service. GatewayRepository drops from ~2629 to ~2375 lines.

Moved to SyncCoordinator:
- setScriptsAndRecord (nativeSetScripts + sync_progress write)
- applyBalancedFilter (I/O wrapper around the pure algorithm)
- maybeReregisterBalanced (cache-aware re-registration check)
- registerAllWalletScripts (multi-wallet script registration)
- balancedFilterAlgorithm (pure partition, top-level function)
- BALANCED_LAG_THRESHOLD + MAX_CONCURRENT_WALLET_SCRIPTS constants
- scriptArgsToWalletId mapping (lock-script args → walletId)
- lastBalancedEligibleSet cache

GatewayRepository keeps:
- The four method shims that delegate (preserving call sites and
  signatures inside the file unchanged for blast-radius safety)
- The makeSyncContext() helper that bundles per-call state
  (activeWalletId, currentNetwork, awaitNodeReady, getWalletSyncBlock,
  onScriptsRegistered) into the SyncContext data class the coordinator
  takes per call
- One read site for scriptArgsToWalletId migrated to
  syncCoordinator.getWalletIdForScript(args)

DI: AppModule provides SyncCoordinator via @Inject constructor + adds
it to the GatewayRepository factory. Existing consumers of
GatewayRepository (every ViewModel + repository test) are unchanged.

Tests: 611/611 pass (4 new in SyncCoordinatorTest covering the moved
algorithm + edge cases). The existing GatewayRepositoryBalancedTest
references balancedFilterAlgorithm via the same package — no import
changes needed.

Phase 2 (DaoOperations extraction) lands next.

Refs #106
